This fixes a bunch of warnings that are in our code, in order for the GYP build to succeed. Review URL: http://codereview.chromium.org/118346
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17865 0039d316-1c4b-4281-b951-d872f2087c98

+++ o3d/converter/cross/converter.cc
@@ -298,7 +298,7 @@ bool Verify(const FilePath& in_filename,
   // shader to the given output file, if there is one.
   if (options.condition && !out_filename.empty()) {
     if (file_util::WriteFile(out_filename, shader_source_in.c_str(),
-                             shader_source_in.size()) == -1) {
+                             static_cast<int>(shader_source_in.size())) == -1) {
       O3D_ERROR(&service_locator) << "Warning: Could not write to output file '"
                                   << FilePathToUTF8(in_filename).c_str() << "'";
     }
